GPS feature has been made.
It now tracks whether are not you're within the area made by a circle.

QR Code scanner now works. Many buttons have been re-arranged.

Added in support for iOS, but there is no development for it due to the lack of a mac and an iOS device. Once aquired development will start.